+++ R/JointAIObject.R
@@ -68,7 +68,7 @@
 #' regression coefficients for each of the main models per design matrix}
 #' \item{\code{par_index_other}}{a list of matrices specifying the indices of
 #' regression coefficients for each covariate model per design matrix}
-#' \item{\code{JAGSmodel}}{The JAGS model as character string.}
+#' \item{\code{jagsmodel}}{The JAGS model as character string.}
 #' \item{\code{mcmc_settings}}{a list containing MCMC sampling related
 #'                             information with elements
 #'      \itemize{