+++ tests/testthat/test-clmm.R
@@ -156,7 +156,7 @@ test_that("models run", {
 })
 
 
-test_that("there are no duplicate betas/alphas in the JAGSmodel", {
+test_that("there are no duplicate betas/alphas in the jagsmodel", {
   expect_null(unlist(lapply(models, find_dupl_parms)))
 })
 
@@ -180,9 +180,9 @@ test_that("data_list remaines the same", {
   print_output(lapply(models, "[[", "data_list"))
 })
 
-test_that("JAGSmodel remaines the same", {
+test_that("jagsmodel remaines the same", {
   skip_on_cran()
-  print_output(lapply(models, "[[", "JAGSmodel"))
+  print_output(lapply(models, "[[", "jagsmodel"))
 })