+++ tests/testthat/test-glm.R
@@ -277,7 +277,7 @@ test_that("models have the correct model and analysis model type", {
 })
 
 
-test_that("there are no duplicate betas/alphas in the JAGSmodel", {
+test_that("there are no duplicate betas/alphas in the jagsmodel", {
   expect_null(unlist(lapply(models, find_dupl_parms)))
 })
 
@@ -301,9 +301,9 @@ test_that("data_list remaines the same", {
   print_output(lapply(models, "[[", "data_list"))
 })
 
-test_that("JAGSmodel remaines the same", {
+test_that("jagsmodel remaines the same", {
   skip_on_cran()
-  print_output(lapply(models, "[[", "JAGSmodel"))
+  print_output(lapply(models, "[[", "jagsmodel"))
 })
 
 test_that("GRcrit and MCerror give same result", {