+++ tests/testthat/test-glmm.R
@@ -294,7 +294,7 @@ if (identical(Sys.getenv("NOT_CRAN"), "true")) {
   })
 
 
-  test_that("there are no duplicate betas/alphas in the JAGSmodel", {
+  test_that("there are no duplicate betas/alphas in the jagsmodel", {
     expect_null(unlist(lapply(models, find_dupl_parms)))
   })